Fixed problem of null sampler state in SamplerStateOp
Problem was, that when creating an invalid sampler state description in
the first evaluation then the output was never set and remained 'null'.
Now in case of an error during sampler state creation it's first checked
if previously a valid sampler state was set (output.value != null). In
that case this is returned. If there was no previously valid sampler
state then the default sampler state is returned.