Fixed resource leak of depth buffers
Problem was that in OutputMergerStage a depth-stencil-view was requested, but
this was never disposed afterwards. So since lately depth buffers and DSVs
were added this turned into a resource leak.
As the OutputMergerStage is currently not using the DSV input anyway, for
now the request of the actually set DSV has been removed.